+++ Sources/StreamChat/Models/UnreadCount.swift
@@ -4,20 +4,8 @@
 
 import Foundation
 
-// TODO: v5: Because we mix the UnreadCount Domain model with Payload we have to set `threads` optional.
-// The reason is because when mapping the unread count from `CurrentUserDTO` we know `threads` can't be null, and will be at least be 0.
-// But since this structure is used to parse unread count from WSEvents, in those events `threads` can be null sometimes, so we have to set it null here.
-
 /// A struct containing information about unread counts of channels and messages.
 public struct UnreadCount: Decodable, Equatable {
-    private enum CodingKeys: String, CodingKey {
-        case currentUserUnreadCount = "me"
-        case userUnreadCount = "user"
-        case channels = "unread_channels"
-        case messages = "total_unread_count"
-        case threads = "unread_threads"
-    }
-
     /// The default value representing no unread channels, messages and threads.
     public static let noUnread = UnreadCount(channels: 0, messages: 0, threads: 0)
 
@@ -28,7 +16,38 @@ public struct UnreadCount: Decodable, Equatable {
     public let messages: Int
 
     /// The number of threads with unread replies if available.
-    public let threads: Int?
+    public let threads: Int
+
+    init(channels: Int, messages: Int, threads: Int) {
+        self.channels = channels
+        self.messages = messages
+        self.threads = threads
+    }
+
+    init(currentUserDTO: CurrentUserDTO) {
+        channels = Int(currentUserDTO.unreadChannelsCount)
+        messages = Int(currentUserDTO.unreadMessagesCount)
+        threads = Int(currentUserDTO.unreadThreadsCount)
+    }
+}
+
+struct UnreadCountPayload: Decodable, Equatable {
+    private enum CodingKeys: String, CodingKey {
+        case currentUserUnreadCount = "me"
+        case userUnreadCount = "user"
+        case channels = "unread_channels"
+        case messages = "total_unread_count"
+        case threads = "unread_threads"
+    }
+
+    /// The number of channels with unread messages.
+    let channels: Int?
+
+    /// The number of unread messages across all channels.
+    let messages: Int?
+
+    /// The number of threads with unread replies if available.
+    let threads: Int?
 
     init(channels: Int, messages: Int, threads: Int?) {
         self.channels = channels
@@ -36,7 +55,7 @@ public struct UnreadCount: Decodable, Equatable {
         self.threads = threads
     }
 
-    public init(from decoder: Decoder) throws {
+    init(from decoder: Decoder) throws {
         let container = try decoder.container(keyedBy: CodingKeys.self)
 
         if let channels = try container.decodeIfPresent(Int.self, forKey: .channels),
@@ -44,14 +63,19 @@ public struct UnreadCount: Decodable, Equatable {
             self.channels = channels
             self.messages = messages
             threads = try container.decodeIfPresent(Int.self, forKey: .threads)
-        } else if let currentUserUnreadCount = try container.decodeIfPresent(UnreadCount.self, forKey: .currentUserUnreadCount) {
+        } else if let currentUserUnreadCount = try container.decodeIfPresent(UnreadCountPayload.self, forKey: .currentUserUnreadCount) {
             channels = currentUserUnreadCount.channels
             messages = currentUserUnreadCount.messages
             threads = currentUserUnreadCount.threads
-        } else if let userUnreadCount = try container.decodeIfPresent(UnreadCount.self, forKey: .userUnreadCount) {
+        } else if let userUnreadCount = try container.decodeIfPresent(UnreadCountPayload.self, forKey: .userUnreadCount) {
             channels = userUnreadCount.channels
             messages = userUnreadCount.messages
             threads = userUnreadCount.threads
+        } else if let threads = try container.decodeIfPresent(Int.self, forKey: .threads) {
+            // For thread parent message new, only unread threads are available.
+            channels = try container.decodeIfPresent(Int.self, forKey: .channels)
+            messages = try container.decodeIfPresent(Int.self, forKey: .messages)
+            self.threads = threads
         } else {
             throw ClientError.EventDecoding(missingValue: "unread_channels, total_unread_count", for: Self.self)
         }