+++ TestTools/StreamChatTestTools/TestData/DummyData/UnreadCount.swift
@@ -5,6 +5,16 @@
 import Foundation
 @testable import StreamChat
 
+extension UnreadCountPayload {
+    static var dummy: Self {
+        .init(
+            channels: Int.random(in: 0...Int.max),
+            messages: Int.random(in: 0...Int.max),
+            threads: Int.random(in: 0...Int.max)
+        )
+    }
+}
+
 extension UnreadCount {
     static var dummy: Self {
         .init(