+++ Tests/StreamChatUITests/SnapshotTests/ChatThreadList/ChatThreadListVC_Tests.swift
@@ -77,6 +77,7 @@ final class ChatThreadListVC_Tests: XCTestCase {
             eventsController: mockedThreadListController.client.eventsController()
         )
         vc.setUpLayout()
+        mockedThreadListController.synchronize_callCount = 0
     }
 
     override func tearDown() {
@@ -133,9 +134,9 @@ final class ChatThreadListVC_Tests: XCTestCase {
         mockedThreadListController.threads_mock = mockThreads
         vc.controller(mockedThreadListController, didChangeState: .initialized)
         vc.controller(mockedThreadListController, didChangeThreads: [])
+        vc.setUpLayout()
         vc.showHeaderBannerView()
         vc.newAvailableThreadIds = [.unique, .unique]
-        vc.setUpLayout()
         AssertSnapshot(vc)
     }
 
@@ -201,6 +202,7 @@ final class ChatThreadListVC_Tests: XCTestCase {
         let newThreadMessageEvent = ThreadMessageNewEvent(
             message: .mock(parentMessageId: .unique),
             channel: .mock(cid: .unique),
+            unreadCount: .noUnread,
             createdAt: .unique
         )
         vc.eventsController(
@@ -218,6 +220,7 @@ final class ChatThreadListVC_Tests: XCTestCase {
         let newThreadMessageEvent = ThreadMessageNewEvent(
             message: .mock(parentMessageId: expectedParentMessageId),
             channel: .mock(cid: .unique),
+            unreadCount: .noUnread,
             createdAt: .unique
         )
         vc.eventsController(
@@ -234,6 +237,7 @@ final class ChatThreadListVC_Tests: XCTestCase {
         let newThreadMessageEvent = ThreadMessageNewEvent(
             message: .mock(parentMessageId: parentMessageId),
             channel: .mock(cid: .unique),
+            unreadCount: .noUnread,
             createdAt: .unique
         )
         try mockedThreadListController.dataStore.database.writeSynchronously { session in
@@ -249,7 +253,6 @@ final class ChatThreadListVC_Tests: XCTestCase {
     // MARK: viewWillAppear
 
     func test_viewWillAppear_whenNewAvailableThreads_shouldRefetchThreadList() {
-        mockedThreadListController.synchronize_callCount = 0
         vc.newAvailableThreadIds = [.unique]
 
         vc.viewWillAppear(false)
@@ -258,11 +261,44 @@ final class ChatThreadListVC_Tests: XCTestCase {
     }
 
     func test_viewWillAppear_whenNoNewThreads_shouldNotRefetchThreadList() {
-        mockedThreadListController.synchronize_callCount = 0
         vc.newAvailableThreadIds = []
 
         vc.viewWillAppear(false)
 
         XCTAssertEqual(mockedThreadListController.synchronize_callCount, 0)
+    }
+
+    // MARK: - didTapOnErrorView
+
+    func test_didTapOnErrorView_shouldShowLoadingView() {
+        mockedThreadListController.threads_mock = []
+        vc.controller(vc.threadListController, didChangeThreads: [])
+
+        vc.didTapOnErrorView()
+
+        XCTAssertEqual(mockedThreadListController.synchronize_callCount, 1)
+        AssertSnapshot(vc, variants: [.defaultLight, .defaultDark])
+    }
+
+    func test_didTapOnErrorView_whenThreadsNotEmpty_shouldShowLoadingHeaderBannerView() {
+        mockedThreadListController.threads_mock = mockThreads
+        vc.controller(vc.threadListController, didChangeThreads: [])
+
+        vc.didTapOnErrorView()
+
+        XCTAssertEqual(mockedThreadListController.synchronize_callCount, 1)
+        AssertSnapshot(vc, variants: [.defaultLight, .defaultDark])
+    }
+
+    // MARK: - didTapOnHeaderBannerView
+
+    func test_didTapOnHeaderBannerView_shouldShowLoadingHeaderBannerView() {
+        mockedThreadListController.threads_mock = mockThreads
+        vc.controller(vc.threadListController, didChangeThreads: [])
+
+        vc.didTapOnHeaderBannerView()
+
+        XCTAssertEqual(mockedThreadListController.synchronize_callCount, 1)
+        AssertSnapshot(vc, variants: [.defaultLight, .defaultDark])
     }
 }